requirements.txt
fastapi[all]
sqlalchemy[asyncio]
asyncpg
aiosqlite
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
uvicorn
pytest
httpx
